major update
* Created a components module
* Simplified the notebook that now
just imports the components
* Improved logging
* Set joblib to use threading
