+++ doodler/__init__.py
@@ -1,12 +1,24 @@
-from .segmentation.annotations_to_segmentations import (
-    label_to_colors,
-)
+import logging
 
-from .segmentation.image_segmentation import (
-    segmentation,
+from .components import (
+    Application,
+    ClassToggleGroup,
+    ComputationSettings,
+    DoodleDrawer,
+    Info,
+    InputImage,
+    Toggle,
 )
 
 __all__ = [
-    'segmentation',
-    'label_to_colors',
+    "Application",
+    "ClassToggleGroup",
+    "ComputationSettings",
+    "DoodleDrawer",
+    "Info",
+    "InputImage",
+    "Toggle",
 ]
+
+# Set default logging handler for a library
+logging.getLogger(__name__).addHandler(logging.NullHandler())
